requirements.txt
pygame
opencv-python
mediapipe
python-chess
